+++ src/main/java/shop/nhnteam04/front/order/controller/OrderController.java
@@ -3,22 +3,14 @@ package shop.nhnteam04.front.order.controller;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import shop.nhnteam04.front.account.user.dto.SecurityUser;
-import shop.nhnteam04.front.order.dto.orders.request.CreateOrderRequest;
-import shop.nhnteam04.front.order.dto.orders.response.OrderDetailResponse;
-import shop.nhnteam04.front.order.dto.orders.response.OrderInfoResponse;
-import shop.nhnteam04.front.order.dto.orders.response.OrderResponse;
 import shop.nhnteam04.front.order.service.OrderService;
 import shop.nhnteam04.front.order.service.PaymentService;
-
-import java.util.List;
 
 @Slf4j
 @Controller
@@ -27,32 +19,6 @@ import java.util.List;
 public class OrderController {
     private final OrderService orderService;
     private final PaymentService paymentService;
-
-    // 유저 주문 내역
-    @GetMapping("/my-list")
-    public ModelAndView userOrderForm(@AuthenticationPrincipal SecurityUser user,
-                                  @RequestParam(defaultValue = "0") int page,
-                                  @RequestParam(defaultValue = "10") int size) {
-        // TODO: html 만들기
-        ModelAndView mav = new ModelAndView("userOrderList");
-
-        Pageable pageable = PageRequest.of(page, size);
-        List<OrderInfoResponse> list = orderService.getUserOrders(user.getId(), pageable);
-        mav.addObject("orders", list);
-
-        return mav;
-    }
-
-    // 주문 상세 정보
-    @GetMapping("/detail")
-    public ModelAndView OrderDetail(@RequestParam("orderId") long orderId) {
-        ModelAndView mav = new ModelAndView("order-detail");
-
-        OrderDetailResponse response = orderService.getOrder(orderId);
-        mav.addObject("orderDetail", response);
-
-        return mav;
-    }
 
     // 주문서 작성
     @GetMapping
@@ -87,6 +53,8 @@ public class OrderController {
         re.addAttribute("orderNumber", orderResponse.getOrderNumber());
         re.addAttribute("finalPrice", orderResponse.getFinalPrice());
         */
+
+        re.addAttribute("orderNumber", "wdawdrfsedfcdsw");
 
         return "redirect:/payment";
     }
